Added "useNext" static method to the Schema
The "useNext " method allows the developer to increase the counter manually and return the new value. It's useful when we want to control when an field should be set to the next available indentity. Aditionally an "autoSet" flag has been added to settings in order to allow the developer to bypass the default save functionality. Also, all the methods now return promises and the callbacks for all functions have been made optional.